requirements.txt
Celery
Celery[Redis]
entsoe-py
requests
requests-mock
delorean